# doc_utils/W2M-Swagger-doc.yaml
openapi: 3.0.1
info:
  title: World 2 Meet
  description: Documentation about the Spaceship service endpoints
  license:
    name: World 2 Meet
    url: http://localhost:3000
  version: 0.1.0
servers:
  - url: http://localhost:3000
    description: Generated server url
paths:
  /spaceship/update/{id}:
    put:
      tags:
        - Spaceship
      summary: Update a Spaceship record
      operationId: updateSpaceship
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SpaceshipForm'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Spaceship'
  /source/update/{id}:
    put:
      tags:
        - Source
      summary: Update an already saved Spaceship record
      operationId: updateSpaceship_1
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SourceUpdateForm'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Source'
  /spaceship/save:
    post:
      tags:
        - Spaceship
      summary: Saved new Spaceship record
      operationId: addNewSpaceship
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SpaceshipForm'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Spaceship'
  /spaceship/{id}:
    get:
      tags:
        - Spaceship
      summary: Get Spaceship by id
      operationId: findById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Spaceship'
  /spaceship/showSpaceships:
    get:
      tags:
        - Spaceship
      summary: Send message request to show all Spaceship images
      operationId: showAllSpaceships
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SpaceshipImgDto'
  /spaceship/byName:
    get:
      tags:
        - Spaceship
      summary: Find Spaceship by name
      operationId: findByName
      parameters:
        - name: name
          in: query
          required: true
          schema:
            type: string
        - name: pageable
          in: query
          required: true
          schema:
            $ref: '#/components/schemas/Pageable'
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Spaceship'
  /spaceship/all:
    get:
      tags:
        - Spaceship
      summary: Get all Spaceship records
      operationId: getAll
      parameters:
        - name: pageable
          in: query
          required: true
          schema:
            $ref: '#/components/schemas/Pageable'
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/PageSpaceship'
  /spaceship/remove/{id}:
    delete:
      tags:
        - Spaceship
      summary: Remove Spaceship by id
      operationId: removeSpaceship
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: string
                enum:
                  - 100 CONTINUE
                  - 101 SWITCHING_PROTOCOLS
                  - 102 PROCESSING
                  - 103 EARLY_HINTS
                  - 103 CHECKPOINT
                  - 200 OK
                  - 201 CREATED
                  - 202 ACCEPTED
                  - 203 NON_AUTHORITATIVE_INFORMATION
                  - 204 NO_CONTENT
                  - 205 RESET_CONTENT
                  - 206 PARTIAL_CONTENT
                  - 207 MULTI_STATUS
                  - 208 ALREADY_REPORTED
                  - 226 IM_USED
                  - 300 MULTIPLE_CHOICES
                  - 301 MOVED_PERMANENTLY
                  - 302 FOUND
                  - 302 MOVED_TEMPORARILY
                  - 303 SEE_OTHER
                  - 304 NOT_MODIFIED
                  - 305 USE_PROXY
                  - 307 TEMPORARY_REDIRECT
                  - 308 PERMANENT_REDIRECT
                  - 400 BAD_REQUEST
                  - 401 UNAUTHORIZED
                  - 402 PAYMENT_REQUIRED
                  - 403 FORBIDDEN
                  - 404 NOT_FOUND
                  - 405 METHOD_NOT_ALLOWED
                  - 406 NOT_ACCEPTABLE
                  - 407 PROXY_AUTHENTICATION_REQUIRED
                  - 408 REQUEST_TIMEOUT
                  - 409 CONFLICT
                  - 410 GONE
                  - 411 LENGTH_REQUIRED
                  - 412 PRECONDITION_FAILED
                  - 413 PAYLOAD_TOO_LARGE
                  - 413 REQUEST_ENTITY_TOO_LARGE
                  - 414 URI_TOO_LONG
                  - 414 REQUEST_URI_TOO_LONG
                  - 415 UNSUPPORTED_MEDIA_TYPE
                  - 416 REQUESTED_RANGE_NOT_SATISFIABLE
                  - 417 EXPECTATION_FAILED
                  - 418 I_AM_A_TEAPOT
                  - 419 INSUFFICIENT_SPACE_ON_RESOURCE
                  - 420 METHOD_FAILURE
                  - 421 DESTINATION_LOCKED
                  - 422 UNPROCESSABLE_ENTITY
                  - 423 LOCKED
                  - 424 FAILED_DEPENDENCY
                  - 425 TOO_EARLY
                  - 426 UPGRADE_REQUIRED
                  - 428 PRECONDITION_REQUIRED
                  - 429 TOO_MANY_REQUESTS
                  - 431 REQUEST_HEADER_FIELDS_TOO_LARGE
                  - 451 UNAVAILABLE_FOR_LEGAL_REASONS
                  - 500 INTERNAL_SERVER_ERROR
                  - 501 NOT_IMPLEMENTED
                  - 502 BAD_GATEWAY
                  - 503 SERVICE_UNAVAILABLE
                  - 504 GATEWAY_TIMEOUT
                  - 505 HTTP_VERSION_NOT_SUPPORTED
                  - 506 VARIANT_ALSO_NEGOTIATES
                  - 507 INSUFFICIENT_STORAGE
                  - 508 LOOP_DETECTED
                  - 509 BANDWIDTH_LIMIT_EXCEEDED
                  - 510 NOT_EXTENDED
                  - 511 NETWORK_AUTHENTICATION_REQUIRED
components:
  schemas:
    SpaceshipForm:
      required:
        - premiereYear
        - sourceName
        - spaceshipName
      type: object
      properties:
        spaceshipName:
          maxLength: 25
          minLength: 3
          type: string
          description: Spaceship reference, name, model, description
          example: Death star
        imageUrl:
          maxLength: 255
          minLength: 10
          type: string
          description: Spaceship image URL
          example: https://shorturl.at/kpNQT
          default: https://shorturl.at/kpNQT
        premiereYear:
          maximum: 9999
          minimum: 1888
          type: integer
          description: Film - series premiere year
          format: int32
          example: 2001
        sourceName:
          maxLength: 25
          minLength: 0
          type: string
          description: Name of, film - series, where the Spaceship appears
          example: Star Wars
        type:
          type: string
          description: Reference to where the Spaceship was seen
          example: SERIES
          default: FILM
          enum:
            - SERIES
            - FILM
            - FILM
            - SERIES
      description: Form to Create or Update the Spaceship entity
    Source:
      type: object
      properties:
        id:
          type: integer
          format: int64
        premiereYear:
          type: integer
          format: int32
        name:
          type: string
        type:
          type: string
          enum:
            - SERIES
            - FILM
    Spaceship:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        source:
          $ref: '#/components/schemas/Source'
        imageUrl:
          type: string
    SourceUpdateForm:
      type: object
      properties:
        premiereYear:
          maximum: 9999
          minimum: 1888
          type: integer
          description: Film - series premiere year
          format: int32
          example: 2001
        name:
          maxLength: 25
          minLength: 0
          type: string
          description: Name of, film - series, where the Spaceship appears
          example: Star Wars
        type:
          type: string
          description: Reference to where the Spaceship was seen
          example: SERIES
          enum:
            - SERIES
            - FILM
            - FILM
            - SERIES
      description: Form to Update the Source entity
    SpaceshipImgDto:
      type: object
      properties:
        name:
          type: string
          description: Spaceship reference, name, model, description
          example: N-1 Starfighter
        imageUrl:
          type: string
          description: Spaceship image URL
          example: https://shorturl.at/jowO4
          default: https://shorturl.at/jowO4
      description: Dto to send some Spaceship info as message request
    Pageable:
      type: object
      properties:
        page:
          minimum: 0
          type: integer
          format: int32
        size:
          minimum: 1
          type: integer
          format: int32
        sort:
          type: array
          items:
            type: string
    PageSpaceship:
      type: object
      properties:
        totalPages:
          type: integer
          format: int32
        totalElements:
          type: integer
          format: int64
        pageable:
          $ref: '#/components/schemas/PageableObject'
        size:
          type: integer
          format: int32
        content:
          type: array
          items:
            $ref: '#/components/schemas/Spaceship'
        number:
          type: integer
          format: int32
        sort:
          type: array
          items:
            $ref: '#/components/schemas/SortObject'
        numberOfElements:
          type: integer
          format: int32
        first:
          type: boolean
        last:
          type: boolean
        empty:
          type: boolean
    PageableObject:
      type: object
      properties:
        pageNumber:
          type: integer
          format: int32
        pageSize:
          type: integer
          format: int32
        offset:
          type: integer
          format: int64
        sort:
          type: array
          items:
            $ref: '#/components/schemas/SortObject'
        paged:
          type: boolean
        unpaged:
          type: boolean
    SortObject:
      type: object
      properties:
        direction:
          type: string
        nullHandling:
          type: string
        ascending:
          type: boolean
        property:
          type: string
        ignoreCase:
          type: boolean